requirements.txt
mysql-connector-python==8.3.0
streamlit==1.31.1
langchain==0.1.8
pandas
python-dotenv
